Fix #294
Forgot to change the URLs which had previously been modified in order to work around Github's caching of embedded images - URLs pointing to external files need to get changed whenever the file is updated, otherwise github will keep using the old cached version. There does not appear to be a way to tell github to not cache images referenced in a .md file and just always load it from the URL specified.