"Fix" #295
The latest chapter in the saga of "arduino" vs "stk500v1" for Arduino as ISP programmer, cribbed from the official core - apparently an admission that changing the protocol to "arduino" for all microcontrollers to make it work with the 32u4 (micro/leo) caused too many problems (some but not all users needing to disable reset on the board with a cap, mostly). So it's split into two programmer menu entries now. Ugh.