Change major version as fix for #313 is breaking change for some
According to Per: "This will be a breaking change for anyone who is using a previously valid complete FQBN for one of the ATTinyCore boards. That would be people using the Arduino IDE's CLI or arduino-cli. Specifying all options in the FQBN was mandatory when using some recent versions of the Arduino IDE and arduino-cli so many will be using them, even if they are just set to the default values."