CI: Only compile for x41 chips using IDE versions that avoid exceeding Windows maximum path length
In Arduino IDE 1.8.5 and older versions, boards with a lot of custom menus can cause the path of the cached core to exceed Windows' MAX_PATH, resulting in a compilation error. To avoid this issue, an #error directive was added that causes compilation to fail when using Arduino IDE < 1.8.6 and ATtiny841 or ATtiny441, thus the Travis CI build must be updated to avoid compiling for these boards with earlier IDE versions.