Don't include reset pin in GPIO count
We don't include the reset pin in the GPIO count. We don't have a menu option for enabling RSTDSBL, so it is confusing to advertise that the part has a GPIO pin that can only be enabled by manually blowing RSTDSBL and leaving the part in a state where it can't be reprogrammed without equipment that most users don't have.