Add (no bootloader) to 2313/4313 line
Just noticed it was missing. Also fixed the capitalization of ATtiny in a few places, and removed the compiler toolchain bug note, as 1.3.x of ATTinyCore works around it, and moved the part specific documentation to the top, because about a third of the inquiries I receive are due to people not reading their part-specific doc page.