Fix pulseIn timeout (#384)
Okay.... FFS, I wish I had thought of checking it this weekend....
because timeout is broken on the official core too, and here I was
trying to figure out how it could be different between official and my
core. And, yes, of course it is fucking broken! It's right there, black
and white, clear as crystal - in the first two loops, all they have to
do is decrement an unsigned long, and there is no need for a separate
comparison because the Z-flag will be set if the last result was zero.
But on the third loop, they have to increment an unsigned long and then
compare it to another unsigned long.
Once you unpick the hideous mess that the compiler output and count the
instructions, the first two loops run 11 clocks and only the final one,
which measures the pulse length, runs 16.

This should fix it... Give this a spin for me will you, and make sure I
haven't broken it for short pulses or something?