Fix #405 - work around arduino-cli bug
arduino-cli incorrectly reports version at 1.6.7, rather than whatever version it is based on. 

This will result in the error message not being shown on IDE 1.6.7 - but if they're using that godawful version of the IDE, they deserve what they get. On 1.6.7 and all versions of arduino-cli which report that version, a #warning is shown instead.