Re-export x7, 828 pinout for size
The x7 image was previously a 7MB png, and the 828 was just over 1MB. (And they didn't compress). That is unreasonably large for a file that will be downloaded by all users of the core. Lowered quality of 828 as far as I could without starting to see artifacts. Cropped white area from x7 pinout, and exported as jpg with the same strategy. The full res images were retained, and I'll figure out some sane way to distribute in the future, that isn't downloaded automatically, especially when board manager installs are so deeply hidden that probably nobody will view the docs there...