Update pins_arduino.h
There was never an obvious choice for default pin on t85. Seeing as the digispark boards are EVERYWHERE, might as well use the one it uses. Could have done with a define for VUSB only, but I don't think there are boards that put it on 0 circulating in the wild.