Support prescaling external clocks
Why not? People always insist on lowering clock speed for battery operation,, and it was easy.

Added some board defs for it. Clarified options that can only be set by burning bootloader

Oh, and added in the tiny88 boards because I just got a pile of them in the mail.

Added HASUSB define for parts where we think USB should work (not used anywhere yet).