Fix (?) timing for 16.5 and 12.8 MHz
This relates to #437. 12.8 should be perfect now, and 16.5 is - i think - within a tenth of a percent. Both were very easy cases (12.8 just needed a hint to work around integer-math'ing, 16.5 is 33/32nds of 16 (which is naturally perfect); timing wants to get 32/33nds adjustment. Can't do that, but we can do 31/32nds which is very, very close.