PWM frequencies
Fixes #463 - for some reason I selected the /32 prescale instead of /8 at low clock speeds (though the #defines just after indicate I had intended to select /8)
Switched pin 1 to timer1 on ATtiny x5
Also now on the parts that can do it, betweeen 3 and 8 MHz we prescale the PWM timers by 32. this means pins 1 and 4 on the t85, all pwm pins 861, and the lonely timer 0 pwm pin on the x7, and get the PWM frequency back towards the 500 Hz ~ 1000 Hz that attinycore tries to do.