Add millis exactness for oddball frequencies
We add a correction to millis() based on preprocessor math.

For power-of-two frequencies, we do nothing since millis(() is exact.
This includes 12.8 MHz due to the roundoff-avoiding integer math.

For other frequencies, we calculate the roundoff error as a fraction
and scale it to a value in [0, 135).  Since 135 = 5 * 3 * 3 * 3 covers
the prime factors of relevant odd clock frequencies, we catch them all.
We use an unsigned char counter in millis() to correct to zero drift.

For arbitrary future frequency choices, the formula will be a close
approximation with relative error 1. / (135 * usecs_per_overflow / 8.),
which should be around 100ppm or a lot less, depending on the numbers.