Correct several cases of duplicate entries, two clocksource option bugs
Turned up by tool by @lerouxb

The 12 MHz external crystal on 461/861 Optiboot was missing it's clock source parameter, because it was set for 8 MHz internal instead. Meaning that if 8 MHz internal was selected, the core would think it was using a crystal. .

the 12 MHz external CLOCK option on 48/88 Optiboot did not work, as the clock source oarameter was being set for the 16 MHz external clock option instead of 12, and not at all on 12. Would likely have generated compile errors.