Fixing incorrect pin to interrupt macros
Changed digitalPinToPCMSKbit to return the bit index instead of
a mask on PCMSK (removed shift).
Fix a potential typo in digitalPinToPCMSK so that it will return
PCMSK1 for port B pins.
Fixed digitalPinToPCICRbit so that PCIE0 is only returned for
pins PB0 to PB3. Use PCIE1 for all other pins.